fixx voice history test error

+++ app/src/main/java/com/mentalys/app/data/remote/response/mental/history/VoiceResponse.kt
@@ -73,7 +73,7 @@ data class VoicePredictionResultResponse(
     val predictedEmotion: String?,
 
     @field:SerializedName("support_percentage")
-    val supportPercentage: Int?,
+    val supportPercentage: Double?,
 
     @field:SerializedName("confidence_scores")
     val confidenceScores: VoiceConfidenceScores?
